refactor: re-export lib constants
