+++ lib/manager/composer/artifacts.ts
@@ -5,7 +5,7 @@ import {
   SYSTEM_INSUFFICIENT_DISK_SPACE,
   TEMPORARY_ERROR,
 } from '../../constants/error-messages';
-import { PlatformId } from '../../constants/platforms';
+import { PlatformId } from '../../constants';
 import * as datasourcePackagist from '../../datasource/packagist';
 import { logger } from '../../logger';
 import { ExecOptions, exec } from '../../util/exec';
