+++ lib/platform/azure/index.ts
@@ -9,7 +9,7 @@ import {
 } from 'azure-devops-node-api/interfaces/GitInterfaces';
 import delay from 'delay';
 import { REPOSITORY_EMPTY } from '../../constants/error-messages';
-import { PlatformId } from '../../constants/platforms';
+import { PlatformId } from '../../constants';
 import { logger } from '../../logger';
 import { BranchStatus, PrState, VulnerabilityAlert } from '../../types';
 import * as git from '../../util/git';
