+++ lib/workers/global/autodiscover.spec.ts
@@ -1,5 +1,5 @@
 import type { RenovateConfig } from '../../config/types';
-import { PlatformId } from '../../constants/platforms';
+import { PlatformId } from '../../constants';
 import * as platform from '../../platform';
 import * as _ghApi from '../../platform/github';
 import * as _hostRules from '../../util/host-rules';
