+++ lib/workers/global/config/parse/env.ts
@@ -2,7 +2,7 @@ import is from '@sindresorhus/is';
 
 import { getOptions } from '../../../../config/options';
 import type { AllConfig, RenovateOptions } from '../../../../config/types';
-import { PlatformId } from '../../../../constants/platforms';
+import { PlatformId } from '../../../../constants';
 import { getDatasourceList } from '../../../../datasource';
 import { logger } from '../../../../logger';
 import type { HostRule } from '../../../../types';
